refactor(services): split scan_progress module and shrink oversized functions
